requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250909
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blake3==1.0.5
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
build==1.3.0
cachetools==6.2.0
catalogue==2.0.10
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
chromadb==1.0.21
click==8.2.1
cloudpathlib==0.22.0
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.9.3
confection==0.1.5
contexttimer==0.3.3
cupy-cuda12x==13.6.0
cymem==2.0.11
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fairscale==0.4.4
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fpdf2==2.8.4
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
google-auth==1.6.3
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.6
humanfriendly==10.0
identify==2.6.14
idna==3.10
imageio==2.37.0
importlib-metadata==8.0.0
importlib-resources==6.5.2
interegular==0.3.3
iopath==0.1.10
ipython==9.5.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
kubernetes==33.1.0
langchain-core==0.3.76
langcodes==3.5.0
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.28
language-data==1.3.0
lark==1.2.2
lazy-loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
marisa-trie==1.3.1
markdown-it-py==4.0.0
markdown2==2.5.4
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.5
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
murmurhash==1.0.13
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.107.3
openai-harmony==0.0.4
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opendatasets==0.1.22
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
orjson==3.11.3
ormsgpack==1.10.0
outlines==0.1.11
outlines-core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
portalocker==3.2.0
posthog==5.4.0
pre-commit==4.3.0
preshed==3.0.10
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-extra-types==2.10.5
pydeck==0.9.1
pygments==2.19.2
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
ray==2.49.1
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.3
streamlit==1.49.1
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.9.9
tiktoken==0.11.0
timm==0.4.12
tokenizers==0.21.4
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.2.0
typer==0.17.4
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
vllm==0.8.5
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
